+++ src/client.js
@@ -5,7 +5,7 @@ import {
 } from 'apollo-cache-inmemory';
 import { HttpLink } from 'apollo-link-http';
 import introspectionQueryResultData from './fragmentTypes.json';
-import config from './config/Config';
+import config from './config';
 
 const fragmentMatcher = new IntrospectionFragmentMatcher({
   introspectionQueryResultData,
