+++ src/pages/TargetPage/Wrapper.js
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import gql from 'graphql-tag';
 import { Skeleton } from '@material-ui/lab';
+import config from '../../config';
 import useBatchDownloader from '../../hooks/useBatchDownloader';
-
-const efoURL =
-  'https://storage.googleapis.com/open-targets-data-releases/alpha-rewrite/static/ontology/diseases_efo.jsonl';
 
 const ASSOCIATIONS_VIZ_QUERY = gql`
   query AssociationsVizQuery(
@@ -50,7 +48,7 @@ function Wrapper({ ensemblId, symbol, Component, aggregationFilters }) {
     () => {
       let isCurrent = true;
       const promises = [
-        fetch(efoURL).then(res => res.text()),
+        fetch(config.efoURL).then(res => res.text()),
         getAllAssociations(),
       ];
       Promise.all(promises).then(data => {
